requirements.txt
pandas
numpy
matplotlib


llama-index
llama-index-llms-dashscope
llama-index-embeddings-dashscope


flask
flask-cors


openai
pydantic
tiktoken
python-dotenv


numpy-stl